+++ core/src/com/gdxtest/TextureRegionBlock.kt
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.gdx.graphics.g2d.TextureRegion
 
 class TextureRegionBlock(val texture: TextureRegion) : IBlock {
-  override fun show(batch: SpriteBatch, x: Int, y: Int) {
-    batch.draw(texture, (x * BLOCK_WIDTH).toFloat(), (y * BLOCK_HEIGHT).toFloat())
+  override fun show(batch: SpriteBatch, x: Float, y: Float) {
+    batch.draw(texture, x * BLOCK_WIDTH, y * BLOCK_HEIGHT)
   }
 }
